feat:nodeのバージョンとquotes設定を変更した

+++ functions/.eslintrc.js
@@ -26,7 +26,7 @@ module.exports = {
     "import",
   ],
   rules: {
-    "quotes": ["error", "double"],
+    "quotes": ["error", "single"],
     "import/no-unresolved": 0,
     "indent": ["error", 2],
   },
